Provide theoretically better heaps for Prim
